Refactor Cuenta model and mappers to enforce validation, set default values, and improve DTO mappings

+++ src/main/java/com/ntt/challenge/dto/CuentaResponseDTO.java
@@ -2,14 +2,15 @@ package com.ntt.challenge.dto;
 
 import com.ntt.challenge.model.TipoCuenta;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CuentaResponseDTO(
         UUID id,
         String numeroCuenta,
         TipoCuenta tipoCuenta,
-        Double saldoInicial,
+        BigDecimal saldoInicial,
         Boolean estado,
-        UUID clienteId,
+        UUID clienteId, // ID del cliente asociado (heredado de la clase Persona)
         String nombreCliente
 ) {}
